add missing EBXL recipes. tmp disable modular armor recipes(Blood Asp) AE2 oredict register fix(Blood Asp) fix depleted mox crashing(Blood Asp) Bump to 5.08.33.3

+++ src/main/java/gregtech/api/items/GT_RadioactiveCellIC_Item.java
@@ -84,16 +84,7 @@ public class GT_RadioactiveCellIC_Item extends GT_RadioactiveCell_Item implement
             }
         }
         if (getDamageOfStack(yourStack) >= getMaxDamageEx() - 1) {
-            // switch (this.numberOfCells) {
-            //       case 1:
-            //reactor.setItemAt(x, y, ItemList.Depleted_Thorium_1.get(1, new Object[0]));
-            //       break;
-            //       case 2:
-            //reactor.setItemAt(x, y, ItemList.Depleted_Thorium_2.get(1, new Object[0]));
-            //       break;
-            //       case 4:
-            // reactor.setItemAt(x, y, ItemList.Depleted_Thorium_4.get(1, new Object[0]));
-            //}
+
                         	reactor.setItemAt(x, y, sDepleted.copy());
         } else if (heatrun) {
             damageItemStack(yourStack, 1);
@@ -103,7 +94,7 @@ public class GT_RadioactiveCellIC_Item extends GT_RadioactiveCell_Item implement
         {
               if (sMox&&reactor.isFluidCooled())
                   {
-                            float breedereffectiveness = reactor.getHeat() / reactor.getMaxHeat();
+                            float breedereffectiveness = (float)reactor.getHeat() / (float)reactor.getMaxHeat();
                 if (breedereffectiveness > 0.5D) {
                       heat *= 2;
                     }
